+++ dwarfeng-datamark-core/src/test/java/com/dwarfeng/datamark/node/example/SingletonQosProcessExample.java
@@ -1,4 +1,4 @@
-package com.dwarfeng.datamark.example;
+package com.dwarfeng.datamark.node.example;
 
 import com.dwarfeng.datamark.stack.service.DatamarkQosService;
 import org.apache.commons.lang3.StringUtils;
@@ -10,18 +10,18 @@ import java.util.List;
 import java.util.Scanner;
 
 /**
- * 多例模式 QoS 流程示例。
+ * 单例模式 QoS 流程示例。
  *
  * @author DwArFeng
  * @since 1.0.0
  */
-public class MultitonQosProcessExample {
+public class SingletonQosProcessExample {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(MultitonQosProcessExample.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(SingletonQosProcessExample.class);
 
     public static void main(String[] args) {
         ClassPathXmlApplicationContext ctx = new ClassPathXmlApplicationContext(
-                "classpath:spring/multiton/application-context*.xml"
+                "classpath:spring/singleton/application-context*.xml"
         );
         ctx.registerShutdownHook();
         ctx.start();